+++ src/util/hex.rs
